c/compute: Prevent moire artefacts in distortion
If distortion shrinks some areas too much, the shader will only sample a
few pixels and cause moire.
Change the shader so that it averages values between samples.
This is similar to what mipmaps would do, except that creating mipmaps
on a compute queue is not trivial, and accessing them in compute shaders
may be complex too.